requirements.txt
torch
torchvision
transformers>=4.44.0
datasets>=2.20.0
peft>=0.12.0
accelerate
timm
gradio
evaluate
numpy
pillow
tqdm
scikit-learn
matplotlib
